Create mockSchema and modularize it
