+++ app.js
@@ -23,27 +23,14 @@ app.use("/", indexRouter);
 app.use("/users", usersRouter);
 
 var graphqlHTTP = require("express-graphql");
-var { buildSchema } = require("graphql");
-
-// Construct a schema, using GraphQL schema language
-var schema = buildSchema(`
-  type Query {
-    hello: String
-  }
-`);
+var { mockSchema, mockResolver } = require("./schemas/mock_api_schema");
 
 // The root provides a resolver function for each API endpoint
-var root = {
-  hello: () => {
-    return "Hello world!";
-  },
-};
-
-app.use(
+var root = app.use(
   "/graphql",
   graphqlHTTP({
-    schema: schema,
-    rootValue: root,
+    schema: mockSchema,
+    rootValue: mockResolver,
     graphiql: true,
   })
 );
